Move libconsumer under common/consumer/
