+++ src/bin/lttng-consumerd/Makefile.am
@@ -8,7 +8,7 @@ lttng_consumerd_SOURCES = lttng-consumerd.c \
 	health-consumerd.c
 
 lttng_consumerd_LDADD = \
-	   $(top_builddir)/src/common/libconsumer.la \
+	   $(top_builddir)/src/common/consumer/libconsumer.la \
 	   $(top_builddir)/src/common/sessiond-comm/libsessiond-comm.la \
 	   $(top_builddir)/src/common/libcommon.la \
 	   $(top_builddir)/src/common/index/libindex.la \
