+++ src/common/consumer/consumer-metadata-cache.c
@@ -30,7 +30,7 @@
 #include <common/utils.h>
 #include <common/sessiond-comm/sessiond-comm.h>
 #include <common/ust-consumer/ust-consumer.h>
-#include <common/consumer.h>
+#include <common/consumer/consumer.h>
 
 #include "consumer-metadata-cache.h"
