+++ src/common/consumer/consumer-stream.c
