+++ src/common/consumer/consumer-stream.h
