+++ src/common/ust-consumer/ust-consumer.h
@@ -22,7 +22,7 @@
 #include <config.h>
 #include <errno.h>
 
-#include <common/consumer.h>
+#include <common/consumer/consumer.h>
 
 #ifdef HAVE_LIBLTTNG_UST_CTL
